Update and rename PciDevice.cs to PCIDevice.cs
